requirements.txt
flask>=2.3.0
openai>=1.0.0
boto3>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
